source/requirements.txt
flask==2.1.3
google-cloud-firestore==2.6.0
google-cloud-logging==3.2.1
google-cloud-storage==2.4.0
grpcio-status==1.47.0
gunicorn==20.1.0
mock-firestore==0.11.0
pip-chill==1.0.1
pytest-cov==3.0.0
